requirements.txt
asgiref==3.9.1
certifi==2025.7.14
charset-normalizer==3.4.2
distlib==0.4.0
Django==5.2.5
django-taggit==6.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.2.2
filelock==3.19.1
idna==3.10
mysql-connector-python==9.4.0
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
PyJWT==2.10.1
requests==2.32.4
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.34.0
